Temporary fix for Matches Warning
The matches plugin would display a warning before if you had no teams in
the DB. You can't blame it for wanting to load something that isn't
there ;)
I'm not confident about the workaround I made for this, however it
should solve the problem by redirecting you to the Addteam page.
Also included in this commit is a small fix for the wording when Adding
a server, I wanted it to be clear that HTML code is what you have to use
currently.