Clansuite Menu Fixed
Previously all the plugin links would display even if you didn't have
them installed. Now the meta file checks to see if the table exists for
each Plugin(aka has the plugin been installed?), if it does, the link
will be displayed.
Also the start of a small todo list...